Fix tests

+++ presage-store-sqlite/src/protocol.rs
@@ -52,7 +52,7 @@ impl SessionStore for SqliteProtocolStore {
         &self,
         address: &ProtocolAddress,
     ) -> Result<Option<SessionRecord>, SignalProtocolError> {
-        let device_id: u32 = address.device_id().into();
+        let device_id: u8 = address.device_id().into();
         let address = address.name();
         query!(
             "SELECT record FROM sessions
@@ -74,7 +74,7 @@ impl SessionStore for SqliteProtocolStore {
         address: &ProtocolAddress,
         record: &SessionRecord,
     ) -> Result<(), SignalProtocolError> {
-        let device_id: u32 = address.device_id().into();
+        let device_id: u8 = address.device_id().into();
         let address = address.name();
         let record = record.serialize()?;
 
@@ -124,7 +124,7 @@ impl SessionStoreExt for SqliteProtocolStore {
         name: &ServiceId,
     ) -> Result<Vec<DeviceId>, SignalProtocolError> {
         let address: String = name.raw_uuid().to_string();
-        let device_id: u32 = (*DEFAULT_DEVICE_ID).into();
+        let device_id: u8 = (*DEFAULT_DEVICE_ID).into();
         query_scalar!(
             "SELECT device_id AS 'id: u32' FROM sessions
             WHERE address = ? AND device_id != ? AND identity = ?",
@@ -145,7 +145,7 @@ impl SessionStoreExt for SqliteProtocolStore {
 
     /// Remove a session record for a recipient ID + device ID tuple.
     async fn delete_session(&self, address: &ProtocolAddress) -> Result<(), SignalProtocolError> {
-        let device_id: u32 = address.device_id().into();
+        let device_id: u8 = address.device_id().into();
         let address = address.name();
         query!(
             "DELETE FROM sessions WHERE address = ? AND device_id = ? AND identity = ?",
@@ -533,7 +533,7 @@ impl IdentityKeyStore for SqliteProtocolStore {
         address: &ProtocolAddress,
         identity: &IdentityKey,
     ) -> Result<IdentityChange, SignalProtocolError> {
-        let device_id: u32 = address.device_id().into();
+        let device_id: u8 = address.device_id().into();
         let address = address.name();
         let bytes = identity.serialize();
 
@@ -607,7 +607,7 @@ impl IdentityKeyStore for SqliteProtocolStore {
         &self,
         address: &ProtocolAddress,
     ) -> Result<Option<IdentityKey>, SignalProtocolError> {
-        let device_id: u32 = address.device_id().into();
+        let device_id: u8 = address.device_id().into();
         let address = address.name();
         query_scalar!(
             "SELECT record FROM identities
@@ -634,7 +634,7 @@ impl SenderKeyStore for SqliteProtocolStore {
         record: &SenderKeyRecord,
     ) -> Result<(), SignalProtocolError> {
         let address = sender.name();
-        let device_id: u32 = sender.device_id().into();
+        let device_id: u8 = sender.device_id().into();
         let record = record.serialize()?;
         query!(
             "INSERT INTO sender_keys
@@ -660,7 +660,7 @@ impl SenderKeyStore for SqliteProtocolStore {
         distribution_id: Uuid,
     ) -> Result<Option<SenderKeyRecord>, SignalProtocolError> {
         let address = sender.name();
-        let device_id: u32 = sender.device_id().into();
+        let device_id: u8 = sender.device_id().into();
         query_scalar!(
             "SELECT record FROM sender_keys
             WHERE address = ? AND device_id = ? AND identity = ? AND distribution_id = ?",
